Change expense.amout's type to decimal/numeric
This avoids rounding errors.

I know that you are not supposed to change migrations, but there was no
data in the expenses  table anyways

+++ api/app_core/src/models/expense.rs
@@ -1,9 +1,11 @@
+use rust_decimal::Decimal;
+
 use super::{user::User, tag::Tag};
 
 pub struct Expense {
     pub expense_id: i32,
     pub user: User,
     pub expense_time: chrono::NaiveDateTime,
-    pub amount: i32,
+    pub amount: Decimal,
     pub tags: Vec<Tag>
 }
